src/gateway-svc/requirements.txt
flask
gunicorn
requests
python-dotenv
Flask-PyMongo
pika
pymon